docs: 补充部分文档
